Added a game timer to attempt a level timer to end game
Either looking to make a level timer to complete in a certain time or
eventually keep record of time for amount of points.  Cannot
test…contacts still not working after so long working on it. Cannot end
game if nothing can touch each other. Tried to use timer to end game
but all it does is pause game with gameOver = true